Add support for tracing unattached calls
These are call instructions with source location corresponding to a
different file. Technically DWARF debug info is supposed to contain each
of the frames/locations corresponding to an inlined call but for
whatever reason gcc isn't preserving this (as far as I can tell). So we
just add a separate command ('X') and let the user add it to the
appropriate line.